move login, beta and home features
